SERVER-80207 Use 4-byte counter for tracking time series bucket direct writes
(cherry picked from commit 04b348823444551b68d87e2ae1f744a4889b1c7c)

+++ src/mongo/db/timeseries/bucket_catalog/bucket_state_registry.h
@@ -79,7 +79,7 @@ enum class BucketState : uint8_t {
  * on the same bucket. Conflicts between multiple simultaneous direct writes are mediated by the
  * storage engine.
  */
-using DirectWriteCounter = std::int8_t;
+using DirectWriteCounter = std::int32_t;
 
 /**
  * A helper struct to hold and synchronize both individual bucket states and global state about the
